requirements.txt
streamlit
langchain-community
youtube-transcript-api
python-dotenv
langchain-anthropic
langchain-openai